requirements.txt
beautifulsoup4==4.9.3
django==3.2.3
requests==2.25.1
psycopg2==2.8.6
python-decouple==3.4